Slim down the recording clients with mockk FetchSpec stubs
The hand-written UnsupportedFetchSpec objects were the bulk of the
recording test clients; return a bare mockk() instead, and inline the
functional-test client into its only test file.
